pyproject.toml
[tool.poetry]
name = "destiny"
version = "0.1.0"
description = ""
authors = ["GiovanyLima <giovanylima.cruz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
black = "22.10.0"
requests = "2.28.1"
python-dotenv = "0.21.0"
requests-toolbelt = "0.10.0"
dataclass-type-validator = "0.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
